refactoring for: - change namespace name: v8cmn => V8Wrap - new namespace NaMacroCommon
