+++ NaMacro/BasicModule.cpp
@@ -1,11 +1,14 @@
 #include "stdafx.h"
 
-#include "v8cmn.h"
+#include "V8Wrap.h"
 #include "BasicModule.h"
+
+#include <Windows.h>
 
 #include <NaLib/NaDebug.h>
 
-#include "Windows.h"
+#include "NaMacroCommon.h"
+
 #include "NaWindow.h"
 #include "NaControl.h"
 #include "NaImage.h"
@@ -89,7 +92,7 @@ void NaBasicModule::Init(Isolate * isolate, Local<ObjectTemplate>& global_templa
 	// Must in Init() not in Create()
 	// (need HandleScope)
 	HandleScope handle_scope(isolate);
-	Local<Object> system_obj = v8cmn::GetSystemObject(isolate);
+	Local<Object> system_obj = V8Wrap::GetSystemObject(isolate);
 
 #define ADD_SYSTEM_METHOD(_js_func)		ADD_OBJ_METHOD(system_obj, _js_func);
 
@@ -192,7 +195,7 @@ void NaBasicModule::method_include(V8_FUNCTION_ARGS)
 		NaString strFullUrl(url.GetFullUrl());
 		NaDebug::Out(L"Include full: %s\n", strFullUrl.wstr());
 
-		script_source = v8cmn::ReadFile(isolate, strFullUrl.cstr());
+		script_source = V8Wrap::ReadFile(isolate, strFullUrl.cstr());
 		script_name = String::NewFromUtf8(isolate, strFullUrl.cstr(), NewStringType::kNormal);
 		if (script_source.IsEmpty())
 		{
@@ -209,8 +212,8 @@ void NaBasicModule::method_include(V8_FUNCTION_ARGS)
 			Script::Compile(context, script_source, &origin).ToLocal(&script);
 			if (script.IsEmpty())
 			{
-				if (v8cmn::g_bReportExceptions)
-					v8cmn::ReportException(isolate, &try_catch);
+				if (V8Wrap::g_bReportExceptions)
+					V8Wrap::ReportException(isolate, &try_catch);
 
 				NaDebug::Out(L"included script is empty!\n");
 				return;
@@ -222,8 +225,8 @@ void NaBasicModule::method_include(V8_FUNCTION_ARGS)
 			script->Run(context);
 			if (try_catch.HasCaught())
 			{
-				if (v8cmn::g_bReportExceptions)
-					v8cmn::ReportException(isolate, &try_catch);
+				if (V8Wrap::g_bReportExceptions)
+					V8Wrap::ReportException(isolate, &try_catch);
 				return;
 			}
 		}
@@ -430,7 +433,7 @@ void NaBasicModule::method_setTimeout(V8_FUNCTION_ARGS)
 // syntax:		exit()
 void NaBasicModule::method_exit(V8_FUNCTION_ARGS)
 {
-	v8cmn::g_bExit = true;
+	NaMacroCommon::g_bExit = true;
 }
 
 // description: pick a window from point
