+++ NaMacro/MouseModule.cpp
@@ -1,7 +1,7 @@
 #include "stdafx.h"
 
 #include "MouseModule.h"
-#include "v8cmn.h"
+#include "V8Wrap.h"
 
 #include "Windows.h"
 
@@ -51,7 +51,7 @@ Local<Object> NaMouseModule::GetMouseObject(Isolate *isolate)
 {
 	// HandleScope 안에서 호출
 
-	Local<Object> system_obj = v8cmn::GetSystemObject(isolate);
+	Local<Object> system_obj = V8Wrap::GetSystemObject(isolate);
 	Local<String> mouse_name = String::NewFromUtf8(isolate, "mouse", NewStringType::kNormal).ToLocalChecked();
 	Local<Value> mouse_value = system_obj->Get(mouse_name);
 	if (!mouse_value.IsEmpty() && mouse_value->IsUndefined())
