+++ NaMacro/NaMacroCommon.h
@@ -1,0 +1,12 @@
+#pragma once
+
+#include <v8.h>
+
+namespace NaMacroCommon {
+	using namespace v8;
+
+	extern bool g_bExit;
+
+	void MessageLoopTillExit(Isolate* isolate);
+
+}
