+++ NaMacro/NaObject.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "v8cmn.h"
+#include "V8Wrap.h"
 
 #include <v8.h>
