+++ NaMacro/ScreenModule.cpp
@@ -44,7 +44,7 @@ Local<Object> NaScreenModule::GetScreenObject(Isolate * isolate)
 {
 	// HandleScope 안에서 호출
 
-	Local<Object> system_obj = v8cmn::GetSystemObject(isolate);
+	Local<Object> system_obj = V8Wrap::GetSystemObject(isolate);
 	Local<String> screen_name = String::NewFromUtf8(isolate, "screen", NewStringType::kNormal).ToLocalChecked();
 	Local<Value> screen_value = system_obj->Get(screen_name);
 	if (!screen_value.IsEmpty() && screen_value->IsUndefined())
